requirements.txt
psycopg2-binary
pandas==1.5.1
packaging==23.2
apache-airflow==2.5.0
apache-airflow-providers-postgres==5.3.1
requests==2.28.1
pydantic==1.10.8
dbt-core==1.7
dbt-postgres==1.7
python-dateutil==2.8.2
cron-converter==1.0.1
protobuf==4.25.3
TA-Lib==0.4.21
openai==0.27.8
tiktoken==0.4.0